NOTYET_FROMLIST: sched/cpufreq: Reduce CFS tasks latency due to DL tasks
Bandwidth required by DEADLINE must always be granted. At the same time,
if there are CFS tasks RUNNABLE, we don't want them to be delayed to
much by DL ones.

However, if (util_cfs + util_rt + util_dl) is not saturated, we don't
consider util_dl for frequency selection but we just add the (minimum)
required bandwidth bw_dl:

   freq = f(util_cfs + util_rt + bw_dl)

This could lead to unwanted latencies for CFS tasks, for example in the
case of a small CFS task being co-scheduler with a low bw_dl but long
period DL task, e.g.

   util_cfs:   5%
   bw_dl:     10%
   period_dl: 1s

In the above case, the DL task can run for 100ms building up to 90%
utilization and still being not enough to saturate the utilization and
jump to the maximum OPP and thus delaying the execution of the CFS task.

To reduce CFS latencies, let's update how DL tasks utilization is
aggregated. When only DL tasks are running we keep using just the dl_bw
contribution. When instead there are also CFS tasks let's always use the
max between dl_bw and dl_util. This will allow to ramp-up the OPP faster
and thus reduce CFS latencies.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>