sched/core: allow sched_setattr() to sleep
Currently, sched_setattr() is called from sys_sched_setattr() under RCU
read lock to ensure a valid reference to the task_struct* it works on.
However, this does not allow to use mutexes in all the following callee.

Let's fix this by using the RCU read lock just to get a task refcount
which is just good enough to grante task_struct* validity for all the
(eventually sleeping) callee.

Suggested-by: Juri Lelli <juri.lelli@redhat.com>
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Suren Baghdasaryan <surenb@google.com>
Cc: Todd Kjos <tkjos@google.com>
Cc: Joel Fernandes <joelaf@google.com>
Cc: Steve Muckle <smuckle@google.com>
Cc: Juri Lelli <juri.lelli@redhat.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-api@vger.kernel.org

---
Changes in v5:
 Message-ID: <20180905104545.GB20267@localhost.localdomain>
 - allow sched_setattr() to sleep
 Others:
 - rebased on v4.19-rc4