arm64: defconfig: Update UFSHCD for Hi3660 soc
Commit 7ee7ef24d02d ("scsi: arm64: defconfig: enable configs for
Hisilicon ufs") enabled the Hisilicon UFS, which means that we can
flash a rootfs to the on-board flash. However, as it stands, the
kernel gets stuck on:

[    3.360733] Waiting for root device /dev/sdd10...

That seems to be because even though we have SCSI_UFS_HISI=y,
SCSI_UFSHCD and SCSI_UFSHCD_PLATFORM are set to 'm', which means the
required drivers won't be built-in.

We need those to load the rootfs and then load the modules, so set
them as built-ins.

Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>