HACK: Revert all the 4.19-rc USB/PD bits to get usb working again
Go back to what we had w/ 4.18, as that worked with the
current patch stack.

This is a terrible hack, and we will need to replace it
with proper usb support with the usptream changes. But that
is currently work in progress, and in the meantime it would
be nice if the board worked.

Signed-off-by: John Stultz <john.stultz@linaro.org>