EXAMPLE: Create frequency/power tuple based EM for hikey960
Use the frequency values of the OPPs for the two CPU types of the
hikey960 and place them into the busy-cost-data sections of the
energy-costs. This creates an Energy Model (EM) with frequency/power
instead of capacity/power tuples.
The core and the cluster busy-cost-data is changed even though the
per-CPU EM only needs the former one.
Finally, the freq-energy-model property is added to each CPU node so
that the EM related device tree parsing code can switch from
capacity/power to frequency/power tuples.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>