EXAMPLE: Add a fake speedbin setup for hikey960
To be able to test the frequency/power tuple enhancement of the legacy
Energy Model (EM) device tree driver a couple of fake speed bin related
frequency/power tuples were added to the busy-cost-data.
The EM of the android-msm-bluecross-4.9 kernel [qcom/sdm845.dtsi] was
taken as an example for the speedbin related setup.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>