tracing: Fix DELCARE_TRACE_CONDITION() re-definitions
The DECLARE_TRACE_CONDITION() macro is pretty much similar to
DECLARE_TRACE() where an additional TP_CONDITION() parameter can be
specified to support trace events filtering based on a condition to be
checked only when the trace point is enabled.

However, the current implementations of DECLARE_TRACE{_CONDITION}() are
misaligned and the "_CONDITION" variant is missing some definitions and
generates compile time warnings and errors for double symbols
definitions.

Fix that by adding the missing definitions. Do that by following the
same template already used for the ! "_CONDITION()" variant.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Steven Rostedt <rostedt@goodmis.org> (maintainer:TRACING)
Cc: Ingo Molnar <mingo@redhat.com> (maintainer:TRACING)
Cc: linux-kernel@vger.kernel.org (open list)