sched/wake_q: Reduce reference counting for special users
Some users, specifically futexes and rwsems, required fixes
that allowed the callers to be safe when wakeups occur before
they are expected by wake_up_q(). Such scenarios also play
games and rely on reference counting, and until now were
pivoting on wake_q doing it. With the wake_q_add() call being
moved down, this can no longer be the case. As such we end up
with a a double task refcounting overhead; and these callers
care enough about this (being rather core-ish).

This patch introduces a wake_q_add_safe() call that serves
for callers that have already done refcounting and therefore the
task is 'safe' from wake_q point of view (int that it requires
reference throughout the entire queue/>wakeup cycle). In the one
case it has internal reference counting, in the other case it
consumes the reference counting.

Cc: liuqi16@baidu.com
Cc: mingo@redhat.com
Cc: Yongji Xie <elohimes@gmail.com>
Cc: Xie Yongji <xieyongji@baidu.com>
Cc: will.deacon@arm.com
Cc: yuanlinsi01@baidu.com
Cc: zhangyu31@baidu.com
Cc: Waiman Long <longman@redhat.com>
Cc: lilin24@baidu.com
Cc: nixun@baidu.com
Cc: andrea.parri@amarulasolutions.com
Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: https://lkml.kernel.org/r/20181218195352.7orq3upiwfdbrdne@linux-r8p5