sched/cpufreq: Make schedutil energy-aware
Schedutil cpufreq governor can benefit from being energy-aware in two
ways:

1) Choose the highest OPP for a given energy cost, allowing to skip
lower frequencies that would not be cheaper in terms of consumed power.

2) In some situations, it can be interesting to spend a little more
power if that can give a useful frequency boost. Energy-awareness allows
to put an upper bound on the extra power that will be used in a
consistent way across platforms.

This commit implements 1) by computing the power used for the minimal
frequency that is needed, and then selects the highest one that consumes
the same power.

Point 2) is implemented by extending the use of an existing heuristic
that checks if there has been some idle time since last call to
schedutil's get_next_freq(). If not, it is assumed the CPU is in a
frequency ramp up phase and will be allowed to spend extra power to
reach a stable OPP in a shorter amount of time.

Signed-off-by: Douglas RAILLARD <douglas.raillard@arm.com>