sched/core: Add bucket local max tracking
Because of bucketization, we potentially end up with:
 - tracking tasks with different task specific clamps in the same bucket
 - use the task specific clamp value only to select a bucket

For example, with 20% bucket size and assuming to have:
  Task1: util_min=25%
  Task2: util_min=35%
we will end up with both tasks being refcounted in the [20..39]% bucket
and always boosting these tasks only to 20%.

In systems where we have only few and well defined clamp values, it
would be useful to track the exact clamp value required by a task
whenever possible.
For example, if a system requires only 23% and 47% boost values
then we can aim at tracking the exact boost required by each task
using only 3 buckets of ~33% size each.

Introduce a mechanism to max aggregate the requested clamp values of
RUNNABLE tasks in the same bucket. Keep it simple by resetting the
bucket value to its base value only when a bucket becomes inactive.
Allows a limited and controlled overboosting margin for tasks recounted
in the same bucket.

In systems where the boost values used are not know in advance, it's
still possible to control the maximum acceptable overboosting margin by
tuning the number of clamp groups. For example, 20 groups ensure a 5%
maximum overboost.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>

--
Changes in v8:
 Message-ID: <20190313193916.GQ2482@worktop.programming.kicks-ass.net>
 - split this code out from the previous patch