cpufreq: Use max|min freq of new policy for scale_[max|min]_freq_capacity()
... instead of max|min freq of policy. In case all little CPUs [0..3]
were hotplugged out and then e.g. CPU3 is hotplugged in again, the
policy->max value is 9999999 in the cpufreq_set_policy() call.
Only the new policy contains the correct value 1766400.

Using 9999999 as policy_max_freq in scale_max_freq_capacity() will set
max_freq_scale of CPU3 to a value much higher than 1024.

This leads to an error in update_cpu_capacity() [kernel/sched/fair.c] in
which the thermal_cap() value of 1024 determines the cpu_capacity_orig
value for the little CPU3. The correct value of 381 will never be set
for CPU3 again.

Change-Id: I8a745ee662a2d2ca22aa97159ee0cc5260ff68d7
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>