sched/core: uclamp: always use enum uclamp_id for clamp_id values
The supported clamp indexes are defined in enum clamp_id however, because of the
code logic in some of the first utilization clamping series version, sometimes
we needed to use unsigned int to represent indexes.

This is not more required since the final version of the uclamp_* APIs can
always use the proper enum uclamp_id type.

Fix it with a bulk rename now that we have all the bits merged.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>