sched/fair: util_est: Add new tracepoints for task and cfs_rq
Each time a task is enqueued/dequeued to/from a rq, the rq estimated utilization
is updated. The estimated utilization of a task is updated instead each time
a task completes an activation.

Add a new pair of tracepoints which allows to have modules defining trace
events to collect utilization estimation updates.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>