sched/fair: Move active balance logic to its own function
The logic to trigger an active balance is already quite lengthy, and
as I'm about to add yet another unlock condition it's probably better
to give it its own corner.

The only annoying requirement is that we need to branch to
out_one_pinned when the active balance is cancelled due to the running
task's affinity. Something like < 0, == 0 and > 0 return values could
suffice, but I went for an enum for clarity.

No change in functionality intended, but do note I had to move the
nr_balance_failed in the same if block as the stop_one_cpu_nowait()
call to not have it on the task pinned path.

Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>