sched/fair: Prevent active LB from preempting higher sched classes
The CFS load balancer can cause the cpu_stopper to run a function to
try and steal a rq's currently running task. However, it so happens
that while only CFS tasks will ever be migrated by that function, we
can end up preempting higher sched class tasks, since it is executed
by the cpu_stopper.

I don't expect this to be exceedingly common: we still need to have
had a busiest group in the first place, which needs

  busiest->sum_nr_running != 0

which is a cfs.h_nr_running sum, but if the remote rq starts executing
an RT/DL task *after* we've collected the stats, we can hit this.

Add some extra checks to not trigger the cpu_stopper if the remote
rq's running task isn't CFS, and to bail out of
actice_load_balance_cpu_stop() when there aren't any CFS tasks to
pull.

Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>