sched/core: Move task requeuing checks into __sched_requeuing_required()
A task dequeue/enqueue cycle is required only when certain changes are
actually required for it. These checks are currently inlined into
__sched_setscheduler() thus contributing in making this function big and
hard to parse.

Move those conditions checks into a dedicated __sched_requeuing_required()
function. This allows also to remove a goto and replace it with a cleaner
if statement.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>  
Cc: Peter Zijlstra <peterz@infradead.org>  
Cc: linux-kernel@vger.kernel.org