sched/core: Move requeing code into __sched_requeue_task()
The task dequeue/enqueue sequence is a well defined pattern we want to use
whenever either scheduler params or task attrs needs to be changed.
This code is currently inline in __sched_setscheduler() which makes it necessary
to allocat locals at the beginning and not easy to track if they are changed
midway this function.

Move the dequeue/enqueue sequence into a dedicated __sched_requeue_task().
This allows also to further simplify __sched_setschuler() by removing some
locals not more required.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>